--- Halcyon.Web/ClientApp/.env
@@ -1,1 +1,0 @@
-BROWSER=none
