--- Halcyon.Web/ClientApp/.env.development
@@ -1,2 +1,0 @@
-PORT=44485
-HTTPS=true
