+++ Halcyon.Web/ClientApp/.eslintignore
@@ -1,2 +1,2 @@
 node_modules
-build
+dist
