+++ Halcyon.Web/ClientApp/.prettierignore
@@ -1,2 +1,2 @@
 node_modules
-build
+dist
