+++ Halcyon.Web/ClientApp/index.html
@@ -43,5 +43,6 @@
     <body>
         <noscript>You need to enable JavaScript to run this app.</noscript>
         <div id="root"></div>
+        <script type="module" src="/src/index.jsx"></script>
     </body>
 </html>
