+++ Halcyon.Web/ClientApp/src/components/CheckboxGroupInput.jsx
