+++ Halcyon.Web/ClientApp/src/components/HasPermission.jsx
