+++ Halcyon.Web/ClientApp/src/components/Hero.jsx
