+++ Halcyon.Web/ClientApp/src/pages/HomePage.jsx
