+++ Halcyon.Web/ClientApp/src/utils/config.js
@@ -1,4 +1,4 @@
 export const config = {
-    VERSION: process.env.REACT_APP_VERSION || '1.0.0',
-    API_URL: process.env.REACT_APP_API_URL || '/api'
+    VERSION: import.meta.env.VITE_VERSION || '1.0.0',
+    API_URL: import.meta.env.VITE_API_URL || '/api'
 };
