+++ Halcyon.Web/ClientApp/vite.config.js
@@ -1,0 +1,76 @@
+import { defineConfig, loadEnv } from 'vite';
+import react from '@vitejs/plugin-react';
+import fs from 'fs';
+import path from 'path';
+import { spawn } from 'child_process';
+
+const generateCerts = () => {
+    const baseFolder =
+        process.env.APPDATA !== undefined && process.env.APPDATA !== ''
+            ? `${process.env.APPDATA}/ASP.NET/https`
+            : `${process.env.HOME}/.aspnet/https`;
+
+    const certificateArg = process.argv
+        .map(arg => arg.match(/--name=(?<value>.+)/i))
+        .filter(Boolean)[0];
+
+    const certificateName = certificateArg
+        ? certificateArg.groups.value
+        : process.env.npm_package_name;
+
+    if (!certificateName) {
+        console.error(
+            'Invalid certificate name. Run this script in the context of an npm/yarn script or pass --name=<<app>> explicitly.'
+        );
+        process.exit(-1);
+    }
+
+    const certFilePath = path.join(baseFolder, `${certificateName}.pem`);
+    const keyFilePath = path.join(baseFolder, `${certificateName}.key`);
+
+    if (!fs.existsSync(certFilePath) || !fs.existsSync(keyFilePath)) {
+        spawn(
+            'dotnet',
+            [
+                'dev-certs',
+                'https',
+                '--export-path',
+                certFilePath,
+                '--format',
+                'Pem',
+                '--no-password'
+            ],
+            { stdio: 'inherit' }
+        ).on('exit', code => process.exit(code));
+    }
+
+    return {
+        cert: certFilePath,
+        key: keyFilePath
+    };
+};
+
+// https://vitejs.dev/config/
+export default defineConfig(({ mode }) => {
+    const env = loadEnv(mode, process.cwd(), '');
+
+    return {
+        plugins: [react()],
+        server: {
+            port: 44485,
+            strictPort: true,
+            https: generateCerts(),
+            proxy: {
+                '^/(api|swagger)/*': {
+                    secure: false,
+                    changeOrigin: true,
+                    target: env.ASPNETCORE_HTTPS_PORT
+                        ? `https://localhost:${env.ASPNETCORE_HTTPS_PORT}`
+                        : env.ASPNETCORE_URLS
+                        ? env.ASPNETCORE_URLS.split(';')[0]
+                        : 'http://localhost:56453'
+                }
+            }
+        }
+    };
+});
